Registration Functionality Completed
All of the registration functionality has been completed. A new user record is inserted into the user table if a user with the inputted email does not already exist. This is after it has passed extensive validation. A new client-side alert function has been added allowing a success, warning, error or info prompt to be displayed to the user.